requirements.txt
asgiref==3.5.2
black==22.6.0
click==8.1.3
colorama==0.4.5
Django==4.0.6
django-extensions==3.2.0
djangorestframework==3.13.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
pytz==2022.1
sqlparse==0.4.2
tomli==2.0.1
tzdata==2022.1
